+++ src/Core/panels/search/index.d.ts
@@ -1,8 +1,9 @@
 import { AbsAmPanel } from "../abs";
 import { AMSuggestion } from "./AMSuggestion";
 export declare class AMSearch extends AbsAmPanel {
-    el_input: HTMLInputElement | null;
-    amSuggestion: AMSuggestion | null;
+    el_input: HTMLInputElement;
+    amSuggestion: AMSuggestion;
+    private interval;
     static factory(p_panel: AbsAmPanel): AMSearch;
     constructor(p_panel: AbsAmPanel);
     init_el(): HTMLElement;